requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
sqlalchemy
psycopg2-binary
langchain
langchain-openai
pinecone
requests
httpx